Add placeholder for login page and refactor config
