+++ hooks/Api/Users/Update/index.ts
@@ -1,4 +1,4 @@
-import CONFIG from '../../../../backend-config';
+import CONFIG from '../../../Authentication/backend-config';
 import { useState, useContext } from 'react';
 import { saveUser } from '../shared';
 import AuthContext from '../../../../context/Authentication';
